+++ test/test-primitive.js
@@ -10,7 +10,7 @@ suite('primitive', function() {
 
 
   test('add "primitive" schema', function() {
-    var err = registry.addSchema({
+    registry.addSchema({
       id: 'primitive',
       type: 'object',
       properties: {
@@ -20,34 +20,27 @@ suite('primitive', function() {
         uint: 123
       }
     });
-
-    assert.strictEqual(undefined, err);
   });
 
 
   test('validate "primitive"', function() {
-    var err = registry.test('primitive', {
+    var errors = registry.test('primitive', {
       v0id: 1,
       flag: false,
       text: 'abc',
       uint: 123
     });
 
-    assert(err);
-    assert(err.validation);
-    assert.deepEqual([
+    assert.deepEqual(errors, [
       ['flag', 'boolean', 'value', false],
       ['v0id', 'null', 'type', 1]
-    ], err.validation);
+    ]);
   });
 
 
   test('remove "primitive" schema', function() {
     assert.deepEqual(registry.getSchemas(), ['primitive']);
-
-    var err = registry.removeSchema('primitive');
-    assert.strictEqual(undefined, err);
-
+    registry.removeSchema('primitive');
     assert.deepEqual([], registry.getSchemas());
   });
 });